Add component usages to characteristics docs on site
