+++ docs/components/index.ts
@@ -1,4 +1,16 @@
 export * from "./AllRenderer";
 export * from "./BackgroundBlock";
+export * from "./BlankBlock";
+export * from "./BorderBlock";
+export * from "./ColorBlock";
+export * from "./CursorBlock";
+export * from "./CornerBlock";
+export * from "./FontSizeBlock";
+export * from "./FontWeightBlock";
+export * from "./LetterSpacingBlock";
+export * from "./LineBlock";
+export * from "./TextBlock";
+export * from "./OutlineBlock";
+export * from "./ShadowBlock";
 export * from "./QAContainer";
 export * from "./QAContainerNoStyleInjection";
