+++ site/docs/foundations/characteristics/accordionView.mdx
@@ -5,4 +5,4 @@ sidebar:
   label: Characteristic types
 ---
 
-<AccordionView />
+<TokenDefinitions />
