+++ site/docs/foundations/characteristics/usage.json
@@ -1,0 +1,161 @@
+{
+  "size": [
+    "Accordion",
+    "Avatar",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FlexLayout",
+    "GridLayout",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Spinner",
+    "StackLayout",
+    "StatusAdornment",
+    "Switch",
+    "Toast",
+    "ToggleButton",
+    "ToggleButtonGroup",
+    "Tooltip"
+  ],
+  "separable": [
+    "Accordion",
+    "FlexLayout",
+    "StackLayout"
+  ],
+  "status": [
+    "Accordion",
+    "Banner",
+    "Checkbox",
+    "Input",
+    "MultilineInput",
+    "RadioButton",
+    "StatusAdornment",
+    "StatusIndicator",
+    "Toast",
+    "Tooltip"
+  ],
+  "spacing": [
+    "Accordion",
+    "Badge",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "MultilineInput",
+    "RadioButton",
+    "StatusAdornment",
+    "Switch",
+    "Toast",
+    "ToggleButton",
+    "ToggleButtonGroup"
+  ],
+  "actionable": [
+    "Accordion",
+    "Button",
+    "ToggleButton"
+  ],
+  "text": [
+    "Accordion",
+    "Avatar",
+    "Badge",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Switch",
+    "Text",
+    "Toast",
+    "ToggleButton",
+    "Tooltip"
+  ],
+  "content": [
+    "Accordion",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "RadioButton",
+    "Switch",
+    "Text",
+    "Tooltip"
+  ],
+  "focused": [
+    "Accordion",
+    "Button",
+    "Card",
+    "Checkbox",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "RadioButton",
+    "Switch",
+    "ToggleButton"
+  ],
+  "duration": [
+    "Accordion",
+    "Card",
+    "Switch"
+  ],
+  "accent": [
+    "Avatar",
+    "Badge",
+    "Card",
+    "Spinner"
+  ],
+  "zIndex": [
+    "Badge",
+    "Checkbox",
+    "RadioButton",
+    "Switch",
+    "Tooltip"
+  ],
+  "container": [
+    "Banner",
+    "Card",
+    "Checkbox",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Switch",
+    "Toast",
+    "ToggleButtonGroup",
+    "Tooltip"
+  ],
+  "overlayable": [
+    "Card",
+    "Toast",
+    "Tooltip"
+  ],
+  "selectable": [
+    "Card",
+    "Checkbox",
+    "RadioButton",
+    "Switch"
+  ],
+  "editable": [
+    "FormField",
+    "Input",
+    "MultilineInput"
+  ],
+  "navigable": [
+    "Link"
+  ],
+  "icon": [
+    "Tooltip"
+  ]
+}
