+++ site/docs/foundations/usage/usage.json
@@ -1,0 +1,382 @@
+{
+  "size": [
+    "Accordion",
+    "Avatar",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FlexLayout",
+    "GridLayout",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Spinner",
+    "StackLayout",
+    "StatusAdornment",
+    "Switch",
+    "Toast",
+    "ToggleButton",
+    "ToggleButtonGroup",
+    "Tooltip",
+    "AppHeader",
+    "Breadcrumbs",
+    "ButtonBar",
+    "CascadingMenu",
+    "ColorChooser",
+    "ComboBoxNext",
+    "ContactDetails",
+    "ContentStatus",
+    "Dialog",
+    "Drawer",
+    "DropdownNext",
+    "FileDropZone",
+    "FormFieldLegacy",
+    "InputLegacy",
+    "LayerLayout",
+    "List",
+    "ListNext",
+    "Logo",
+    "MenuButton",
+    "Metric",
+    "NavigationItem",
+    "Overlay",
+    "Pagination",
+    "Pill",
+    "PillNext",
+    "CircularProgress",
+    "LinearProgress",
+    "QueryInput",
+    "SkipLink",
+    "Slider",
+    "TrackerConnector",
+    "TrackerStep",
+    "Tabs",
+    "TokenizedInput",
+    "Toolbar",
+    "OverflowPanel",
+    "Tree"
+  ],
+  "separable": [
+    "Accordion",
+    "FlexLayout",
+    "StackLayout",
+    "AppHeader",
+    "CascadingMenu",
+    "ContactDetails",
+    "Logo",
+    "QueryInput"
+  ],
+  "status": [
+    "Accordion",
+    "Banner",
+    "Checkbox",
+    "Input",
+    "MultilineInput",
+    "RadioButton",
+    "StatusAdornment",
+    "StatusIndicator",
+    "Toast",
+    "Tooltip",
+    "ColorChooser",
+    "Dialog",
+    "FileDropZone",
+    "FormFieldLegacy",
+    "Metric",
+    "TrackerStep"
+  ],
+  "spacing": [
+    "Accordion",
+    "Badge",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "MultilineInput",
+    "RadioButton",
+    "StatusAdornment",
+    "Switch",
+    "Toast",
+    "ToggleButton",
+    "ToggleButtonGroup",
+    "ComboBoxNext",
+    "Dialog",
+    "Drawer",
+    "DropdownNext",
+    "Logo",
+    "NavigationItem",
+    "PillNext",
+    "LinearProgress",
+    "TrackerConnector",
+    "TrackerStep",
+    "ToastGroup"
+  ],
+  "actionable": [
+    "Accordion",
+    "Button",
+    "ToggleButton",
+    "Dropdown",
+    "MenuButton",
+    "Pill",
+    "PillNext",
+    "SkipLink",
+    "Tabs",
+    "TokenizedInput",
+    "Toolbar",
+    "OverflowPanel"
+  ],
+  "text": [
+    "Accordion",
+    "Avatar",
+    "Badge",
+    "Banner",
+    "Button",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Switch",
+    "Text",
+    "Toast",
+    "ToggleButton",
+    "Tooltip",
+    "ColorChooser",
+    "ComboBoxNext",
+    "ContentStatus",
+    "Dialog",
+    "Dropdown",
+    "DropdownNext",
+    "EditableLabel",
+    "FileDropZone",
+    "FormFieldLegacy",
+    "InputLegacy",
+    "List",
+    "ListNext",
+    "Metric",
+    "NavigationItem",
+    "Overlay",
+    "Pagination",
+    "Pill",
+    "PillNext",
+    "CircularProgress",
+    "LinearProgress",
+    "QueryInput",
+    "SkipLink",
+    "Slider",
+    "TrackerStep",
+    "Tabs",
+    "Toolbar",
+    "Tree"
+  ],
+  "content": [
+    "Accordion",
+    "Card",
+    "Checkbox",
+    "FormField",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "RadioButton",
+    "Switch",
+    "Text",
+    "Tooltip",
+    "ContactDetails",
+    "ContentStatus",
+    "Dialog",
+    "DropdownNext",
+    "FileDropZone",
+    "FormattedInput",
+    "InputLegacy",
+    "List",
+    "Logo",
+    "Metric",
+    "NavigationItem",
+    "Overlay",
+    "CircularProgress",
+    "LinearProgress",
+    "QueryInput",
+    "SkipLink",
+    "Slider",
+    "TrackerStep",
+    "OverflowPanel",
+    "Tree"
+  ],
+  "focused": [
+    "Accordion",
+    "Button",
+    "Card",
+    "Checkbox",
+    "Input",
+    "Link",
+    "MultilineInput",
+    "RadioButton",
+    "Switch",
+    "ToggleButton",
+    "ComboBoxNext",
+    "ContactDetails",
+    "DropdownNext",
+    "FormFieldLegacy",
+    "InputLegacy",
+    "List",
+    "NavigationItem",
+    "Pill",
+    "PillNext",
+    "SkipLink",
+    "Slider",
+    "TrackerStep",
+    "TokenizedInput"
+  ],
+  "duration": [
+    "Accordion",
+    "Card",
+    "Switch",
+    "NavigationItem"
+  ],
+  "accent": [
+    "Avatar",
+    "Badge",
+    "Card",
+    "Spinner",
+    "ContactDetails",
+    "Dialog",
+    "CircularProgress",
+    "LinearProgress",
+    "Slider"
+  ],
+  "zIndex": [
+    "Badge",
+    "Checkbox",
+    "RadioButton",
+    "Switch",
+    "Tooltip",
+    "ComboBoxNext",
+    "Dialog",
+    "Drawer",
+    "Dropdown",
+    "DropdownNext",
+    "EditableLabel",
+    "LayerLayout",
+    "Overlay",
+    "PillNext",
+    "QueryInput",
+    "Scrim",
+    "ToastGroup"
+  ],
+  "container": [
+    "Banner",
+    "Card",
+    "Checkbox",
+    "MultilineInput",
+    "Panel",
+    "RadioButton",
+    "Switch",
+    "Toast",
+    "ToggleButtonGroup",
+    "Tooltip",
+    "AppHeader",
+    "CascadingMenu",
+    "Dialog",
+    "Drawer",
+    "Dropdown",
+    "FileDropZone",
+    "LayerLayout",
+    "List",
+    "ListNext",
+    "Overlay",
+    "QueryInput",
+    "Toolbar",
+    "OverflowPanel"
+  ],
+  "overlayable": [
+    "Card",
+    "Toast",
+    "Tooltip",
+    "AppHeader",
+    "CascadingMenu",
+    "ComboBoxNext",
+    "Dialog",
+    "Drawer",
+    "DropdownNext",
+    "LayerLayout",
+    "Overlay",
+    "Scrim",
+    "SkipLink",
+    "Tabs",
+    "OverflowPanel"
+  ],
+  "selectable": [
+    "Card",
+    "Checkbox",
+    "RadioButton",
+    "Switch",
+    "CascadingMenu",
+    "ComboBoxNext",
+    "ContactDetails",
+    "DropdownNext",
+    "List",
+    "Pagination",
+    "Pill",
+    "PillNext",
+    "Tabs",
+    "TokenizedInput",
+    "OverflowPanel",
+    "Tree"
+  ],
+  "editable": [
+    "FormField",
+    "Input",
+    "MultilineInput",
+    "ColorChooser",
+    "DropdownNext",
+    "FormFieldLegacy",
+    "InputLegacy",
+    "TokenizedInput"
+  ],
+  "navigable": [
+    "Link",
+    "NavigationItem",
+    "Tabs"
+  ],
+  "icon": [
+    "Tooltip"
+  ],
+  "animation": [
+    "DeckItem",
+    "DeckLayout",
+    "Dialog",
+    "Drawer",
+    "LayerLayout",
+    "ParentChildItem"
+  ],
+  "target": [
+    "FileDropZone"
+  ],
+  "palette": [
+    "FormFieldLegacy",
+    "Pagination",
+    "Pill",
+    "CircularProgress",
+    "LinearProgress",
+    "Slider",
+    "TrackerConnector",
+    "TokenizedInput"
+  ],
+  "color": [
+    "List"
+  ],
+  "track": [
+    "CircularProgress",
+    "TrackerConnector"
+  ],
+  "draggable": [
+    "Tabs"
+  ]
+}
